+++ CHANGELOG.md
@@ -8,6 +8,10 @@ New features:
   `dynPunchTimesheetersNotice` drives a new in-HRS-app message (presented just
   like `notification`) to employees with the new `ROLE_UW_DYN_AM_PUNCH_TIME`
   role. ([HRSPLT-346][])
++ In Time and Absence, a new optional `portlet-preference`
+  `nonDynPunchTimesheetersNotice` drives a new in-HRS-app message (presented
+  just like `notification`) to employees who see the Timesheet button but who
+  do not have the new `ROLE_UW_DYN_AM_PUNCH_TIME` role. ([HRSPLT-346][])
 
 Changes: